Re-combine (some) classes into single file
We need to refactor our classes to avoid circular dependencies.
